Add Python 3.x support; maintain Python 2.x compatibility
There are two issues here:

    * encoding/decoding of bytes/str (3.x) and str/unicode (2.x)
    * module renaming

The first problem arises because Python 3 forces us to be more explicit
about character encodings. This is a Good Thing, because otherwise, we
are left as defenseless, or Just Plain Wrong.

The second is really just a "janitorial" nuisance: Rename the offending
modules, and all references, and we are good to go.