New method Client.setConnectionProperties()
We want a special, santioned way to pass ConnectionProperties to a
Client, rather than kludging it into the __init__() args. We also gain a
way to reset the connection properties later, or to delay setting them
innitially.

Add Client.setconnectionproperties(), which sets (or resets) the default
headers to use for future requests.

Do not modify Client._default_headers. Instead, use a new
Client.default_headers variable; this allows setConnectionProperties to
be idempotent, starting from the same dict of default headers each time.