New class API: Toplevel API object
This class is the superclass for Github (and any new REST clients we may
invent). It add the setClient and setConnectionProperties methods (to
Github) to solidify the external API for setting up these things. This
makes it much easier to create new subclasses, since the __init__() can
consist of merely 2 method calls.